feat: :sparkles: set create/update licenses to private
